Extract table classes, add ColorGradients table
