+++ lib/data/local/tables/transaction_categories.dart
@@ -1,0 +1,8 @@
+import 'package:drift/drift.dart';
+
+class TransactionCategories extends Table {
+  IntColumn get id => integer().autoIncrement()();
+  TextColumn get name => text()();
+  DateTimeColumn get createdAt => dateTime().withDefault(currentDateAndTime)();
+  DateTimeColumn get updatedAt => dateTime()();
+}
